spotify/requirements.txt
absl-py==1.1.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
dm-tree==0.1.8
docker-pycreds==0.4.0
flatbuffers==1.12
flax==0.5.2
fonttools==4.39.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.7
jaxlib==0.4.7
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
msgpack==1.0.5
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.2
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
promise==2.3
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.9
scipy==1.10.1
sentry-sdk==1.21.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
toolz==0.12.0
typed-ast==1.5.4
typing-extensions==4.2.0
urllib3==2.0.1
wandb==0.13.4
Werkzeug==2.3.3
wrapt==1.15.0
zipp==3.15.0
